portal/pyproject.toml
[tool.poetry]
name = "portal"
version = "0.0.0"  # DO NOT CHANGE -- set during build
description = "Build and deploy functions."
authors = ["Paul Lockaby <paul@paullockaby.com>"]
packages = [{include = "lmnop", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
tenacity = "^8.0.1"
psycopg2-binary = "^2.9.2"
Flask = "^2.0.2"
Flask-WTF = "^0.15.1"
requests = "^2.26.0"
zxcvbn = "^4.4.28"
passlib = "^1.7.4"
gunicorn = "^20.1.0"
bcrypt = "^3.2.0"
gevent = "^21.8.0"
importlib-resources = "^5.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-flake8 = "^1.0.7"
flake8 = "^3.9.2"
flake8-bandit = "^2.1.2"
flake8-builtins = "^1.5.3"
flake8-bugbear = "^21.11.29"
flake8-comprehensions = "^3.7.0"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--flake8 -p no:warnings --log-level=ERROR"
